+++ spec/slack-ruby-bot/commands/about_spec.rb
@@ -10,6 +10,9 @@ describe SlackRubyBot::Commands::Default do
   it 'upcase' do
     expect(message: SlackRubyBot.config.user.upcase).to respond_with_slack_message(SlackRubyBot::ABOUT)
   end
+  it 'name:' do
+    expect(message: "#{SlackRubyBot.config.user}:").to respond_with_slack_message(SlackRubyBot::ABOUT)
+  end
   it 'id' do
     expect(message: "<@#{SlackRubyBot.config.user_id}>").to respond_with_slack_message(SlackRubyBot::ABOUT)
   end