+++ spec/slack-ruby-bot/commands/hi_spec.rb
@@ -7,6 +7,9 @@ describe SlackRubyBot::Commands::Hi do
   it 'says hi' do
     expect(message: "#{SlackRubyBot.config.user} hi").to respond_with_slack_message('Hi <@user>!')
   end
+  it 'says hi to bot:' do
+    expect(message: "#{SlackRubyBot.config.user}: hi").to respond_with_slack_message('Hi <@user>!')
+  end
   it 'says hi to @bot' do
     expect(message: "<@#{SlackRubyBot.config.user_id}> hi").to respond_with_slack_message('Hi <@user>!')
   end